Fixing seed type

+++ nextroute/solver.go
@@ -76,7 +76,7 @@ type SolveOptions struct {
 	Iterations        int           `json:"iterations"  usage:"number of iterations"`
 	MaximumDuration   time.Duration `json:"maximum_duration"  usage:"maximum duration of solver in seconds"`
 	RestartIterations int           `json:"restart_iterations"  usage:"number of iterations before restart"`
-	Seed              int           `json:"seed"  usage:"seed for random generator"`
+	Seed              int64         `json:"seed"  usage:"seed for random generator"`
 }
 
 // Solver is the interface for a solver.
